change data type as unint8_t

+++ src/csrc/mat.c
@@ -8,15 +8,15 @@
 #include <stdio.h>
 #include "mat.h"
 
-float **MatRotate180(float **mat, MatSize mat_size)
+uint8_t **MatRotate180(uint8_t **mat, MatSize mat_size)
 {
 	int column, row;
 	int outSizeW = mat_size.columns;
 	int outSizeH = mat_size.rows;
-	float **outputData = (float **)malloc(outSizeH * sizeof(float *));
+	uint8_t **outputData = (uint8_t **)malloc(outSizeH * sizeof(uint8_t *));
 
 	for (int i = 0; i < outSizeH; i++)
-		outputData[i] = (float *)malloc(outSizeW * sizeof(float));
+		outputData[i] = (uint8_t *)malloc(outSizeW * sizeof(uint8_t));
 
 	for (row = 0; row < outSizeH; row++)
 		for (column = 0; column < outSizeW; column++)
@@ -25,7 +25,7 @@ float **MatRotate180(float **mat, MatSize mat_size)
 	return outputData;
 }
 
-float **MatCorrelation(float **map, MatSize map_size, float **inputData, MatSize inSize, int type)
+uint8_t **MatCorrelation(uint8_t **map, MatSize map_size, uint8_t **inputData, MatSize inSize, int type)
 {
 	int halfmapsizew;
 	int halfmapsizeh;
@@ -43,12 +43,12 @@ float **MatCorrelation(float **map, MatSize map_size, float **inputData, MatSize
 	//
 	int outSizeW = inSize.columns + (map_size.columns - 1); //
 	int outSizeH = inSize.rows + (map_size.rows - 1);
-	float **outputData = (float **)malloc(outSizeH * sizeof(float *)); //
+	uint8_t **outputData = (uint8_t **)malloc(outSizeH * sizeof(uint8_t *)); //
 	for (int i = 0; i < outSizeH; i++)
-		outputData[i] = (float *)calloc(outSizeW, sizeof(float));
+		outputData[i] = (uint8_t *)calloc(outSizeW, sizeof(uint8_t));
 
 	//
-	float **exInputData = MatEdgeExpand(inputData, inSize, map_size.columns - 1, map_size.rows - 1);
+	uint8_t **exInputData = MatEdgeExpand(inputData, inSize, map_size.columns - 1, map_size.rows - 1);
 
 	for (int j = 0; j < outSizeH; j++)
 		for (int i = 0; i < outSizeW; i++)
@@ -69,7 +69,7 @@ float **MatCorrelation(float **map, MatSize map_size, float **inputData, MatSize
 		return outputData;
 	case SAME:
 	{
-		float **sameres = MatEdgeShrink(outputData, outSize, halfmapsizew, halfmapsizeh);
+		uint8_t **sameres = MatEdgeShrink(outputData, outSize, halfmapsizew, halfmapsizeh);
 		for (int i = 0; i < outSize.rows; i++)
 			free(outputData[i]);
 		free(outputData);
@@ -77,7 +77,7 @@ float **MatCorrelation(float **map, MatSize map_size, float **inputData, MatSize
 	}
 	case VALID:
 	{
-		float **validres;
+		uint8_t **validres;
 		if (map_size.rows % 2 == 0 && map_size.columns % 2 == 0)
 			validres = MatEdgeShrink(outputData, outSize, halfmapsizew * 2 - 1, halfmapsizeh * 2 - 1);
 		else
@@ -92,11 +92,11 @@ float **MatCorrelation(float **map, MatSize map_size, float **inputData, MatSize
 	}
 }
 
-float **MatCov(float **map, MatSize map_size, float **inputData, MatSize inSize, int type)
+uint8_t **MatCov(uint8_t **map, MatSize map_size, uint8_t **inputData, MatSize inSize, int type)
 {
 	/*Convolution array in 2D*/
-	float **flipmap = MatRotate180(map, map_size);
-	float **res = MatCorrelation(flipmap, map_size, inputData, inSize, type);
+	uint8_t **flipmap = MatRotate180(map, map_size);
+	uint8_t **res = MatCorrelation(flipmap, map_size, inputData, inSize, type);
 	int i;
 	for (i = 0; i < map_size.rows; i++)
 		free(flipmap[i]);
@@ -104,14 +104,14 @@ float **MatCov(float **map, MatSize map_size, float **inputData, MatSize inSize,
 	return res;
 }
 
-float **MatUpSample(float **mat, MatSize mat_size, int upc, int upr)
+uint8_t **MatUpSample(uint8_t **mat, MatSize mat_size, int upc, int upr)
 {
 	int i, j, m, n;
 	int c = mat_size.columns;
 	int r = mat_size.rows;
-	float **res = (float **)malloc((r * upr) * sizeof(float *));
+	uint8_t **res = (uint8_t **)malloc((r * upr) * sizeof(uint8_t *));
 	for (i = 0; i < (r * upr); i++)
-		res[i] = (float *)malloc((c * upc) * sizeof(float));
+		res[i] = (uint8_t *)malloc((c * upc) * sizeof(uint8_t));
 
 	for (j = 0; j < r * upr; j = j + upr)
 	{
@@ -126,21 +126,21 @@ float **MatUpSample(float **mat, MatSize mat_size, int upc, int upr)
 	return res;
 }
 
-float **MatEdgeExpand(float **mat, MatSize mat_size, int addc, int addr)
+uint8_t **MatEdgeExpand(uint8_t **mat, MatSize mat_size, int addc, int addr)
 {
 	int i, j;
 	int c = mat_size.columns;
 	int r = mat_size.rows;
-	float **res = (float **)malloc((r + 2 * addr) * sizeof(float *));
+	uint8_t **res = (uint8_t **)malloc((r + 2 * addr) * sizeof(uint8_t *));
 	for (i = 0; i < (r + 2 * addr); i++)
-		res[i] = (float *)malloc((c + 2 * addc) * sizeof(float));
+		res[i] = (uint8_t *)malloc((c + 2 * addc) * sizeof(uint8_t));
 
 	for (j = 0; j < r + 2 * addr; j++)
 	{
 		for (i = 0; i < c + 2 * addc; i++)
 		{
 			if (j < addr || i < addc || j >= (r + addr) || i >= (c + addc))
-				res[j][i] = (float)0.0;
+				res[j][i] = (uint8_t)0.0;
 			else
 				res[j][i] = mat[j - addr][i - addc];
 		}
@@ -148,14 +148,14 @@ float **MatEdgeExpand(float **mat, MatSize mat_size, int addc, int addr)
 	return res;
 }
 
-float **MatEdgeShrink(float **mat, MatSize mat_size, int shrinkc, int shrinkr)
+uint8_t **MatEdgeShrink(uint8_t **mat, MatSize mat_size, int shrinkc, int shrinkr)
 {
 	int i, j;
 	int c = mat_size.columns;
 	int r = mat_size.rows;
-	float **res = (float **)malloc((r - 2 * shrinkr) * sizeof(float *));
+	uint8_t **res = (uint8_t **)malloc((r - 2 * shrinkr) * sizeof(uint8_t *));
 	for (i = 0; i < (r - 2 * shrinkr); i++)
-		res[i] = (float *)malloc((c - 2 * shrinkc) * sizeof(float));
+		res[i] = (uint8_t *)malloc((c - 2 * shrinkc) * sizeof(uint8_t));
 
 	for (j = 0; j < r; j++)
 	{
@@ -168,7 +168,7 @@ float **MatEdgeShrink(float **mat, MatSize mat_size, int shrinkc, int shrinkr)
 	return res;
 }
 
-void MatSaving(float **mat, MatSize mat_size, const char *filename)
+void MatSaving(uint8_t **mat, MatSize mat_size, const char *filename)
 {
 	FILE *fp = NULL;
 	fp = fopen(filename, "wb");
@@ -177,11 +177,11 @@ void MatSaving(float **mat, MatSize mat_size, const char *filename)
 
 	int i;
 	for (i = 0; i < mat_size.rows; i++)
-		fwrite(mat[i], sizeof(float), mat_size.columns, fp);
+		fwrite(mat[i], sizeof(uint8_t), mat_size.columns, fp);
 	fclose(fp);
 }
 
-void MatAdd(float **res, float **mat1, MatSize mat_size1, float **mat2, MatSize mat_size2)
+void MatAdd(uint8_t **res, uint8_t **mat1, MatSize mat_size1, uint8_t **mat2, MatSize mat_size2)
 {
 	/*accumulation from 2D to 1D*/
 	int i, j;
@@ -193,7 +193,7 @@ void MatAdd(float **res, float **mat1, MatSize mat_size1, float **mat2, MatSize 
 			res[i][j] = mat1[i][j] + mat2[i][j];
 }
 
-void MatMultifactor(float **res, float **mat, MatSize mat_size, float factor)
+void MatMultifactor(uint8_t **res, uint8_t **mat, MatSize mat_size, uint8_t factor)
 {
 	int i, j;
 	for (i = 0; i < mat_size.rows; i++)
@@ -201,9 +201,9 @@ void MatMultifactor(float **res, float **mat, MatSize mat_size, float factor)
 			res[i][j] = mat[i][j] * factor;
 }
 
-float MatSum(float **mat, MatSize mat_size)
+uint8_t MatSum(uint8_t **mat, MatSize mat_size)
 {
-	float sum = 0.0;
+	uint8_t sum = 0.0;
 	int i, j;
 	for (i = 0; i < mat_size.rows; i++)
 		for (j = 0; j < mat_size.columns; j++)
