+++ src/headers/mnist.h
@@ -4,11 +4,11 @@
 
 #ifndef MNIST_H_
 #define MNIST_H_
-
+#include <stdint.h>
 typedef struct MnistImage{  //Structrue for Mnist Image.
 	int number_of_columns;             //Define the width of image. --- columns
 	int number_of_rows;                //Define the highth ot image. --- rows
-	float** image_data;      //Dynamic Define 2D Image Data Arrary. 
+	uint8_t** image_data;      //Dynamic Define 2D Image Data Arrary. 
 }MnistImage;
 
 typedef struct MnistImageArray{ //Structrue for Mnist Images Array
@@ -18,7 +18,7 @@ typedef struct MnistImageArray{ //Structrue for Mnist Images Array
 
 typedef struct MnistLabel{  //Strutrue for Mnist Label
 	int label_length;                    //The length of Label 
-	float* LabelData;         //The data of label
+	uint8_t* LabelData;         //The data of label
 }MnistLabel;
 
 typedef struct MnistLabelArray{ //Structrue for Mnist Label Arrary
